test: 검색 예산 스위트를 갈라 낸다 (#75)
HwpSearchControllerTests 가 type_body_length(400줄, error)를 넘겨 CI lint 가
막힌다. 매치 상한과 캐시 축출을 HwpSearchBudgetTests 로 옮겼다 — 나누는 축은
AGENTS.md 가 "예산 (성능 — 결과의 의미를 바꾸지 않는다)"으로 묶어 둔 그
경계이고, 픽스처 헬퍼는 네이티브 스위트들처럼 각자 들고 있다.
